refactor devided views tests
